refactor(core): extract scp-identity as independent workspace crate (closes #93)
Move the identity subsystem (~4,891 lines across 6 files) from
scp-core/src/identity/ into a standalone scp-identity workspace crate.
The identity module has zero reverse dependencies on other scp-core
modules, making this a clean extraction that enforces the boundary at
the compiler level.

What moved:
- identity/mod.rs -> scp-identity/src/lib.rs
- identity/{dht,dht_client,document,republish,cache}.rs -> scp-identity/src/

What changed:
- scp-identity depends only on scp-platform and external crates
- scp-core re-exports scp-identity (no breaking change for consumers)
- cache.rs SystemClock inlines now_secs() to avoid scp-core dependency
- All 113 identity tests run in isolation via cargo test -p scp-identity
- All 2496+ workspace tests continue to pass
